Fix test

+++ enterprise/cmd/batcheshelper/log/log_test.go
@@ -33,7 +33,7 @@ func TestLogger_WriteEvent(t *testing.T) {
 
 	assert.Equal(t, "TASK_STEP", actual["operation"])
 	assert.Equal(t, "STARTED", actual["status"])
-	assert.Equal(t, 1.0, actual["metadata"].(map[string]interface{})["step"])
+	assert.Equal(t, float64(1), actual["metadata"].(map[string]interface{})["step"])
 	assert.Equal(t, "BAR", actual["metadata"].(map[string]interface{})["env"].(map[string]interface{})["FOO"])
 	assert.Regexp(t, `^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z$`, actual["timestamp"])
 }
